pyproject.toml
[project]
name = "materializer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.1.2",
    "pillow>=11.0.0",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
]
